### 2019-11-08 #1876 @whedon
[PRE REVIEW]: Plonk: Smoothed particle hydrodynamics analysis and visualization with Python
**Submitting author:** @dmentipl (<a href="http://orcid.org/0000-0002-5526-8798">Daniel Mentiplay</a>)
**Repository:** <a href="https://github.com/dmentipl/plonk" target ="_blank">https://github.com/dmentipl/plonk</a>
**Version:** v0.2.0
