### 2019-11-08 #1876 @dfm
@dmentipl: please take a look at [this list](https://bitly.com/joss-reviewers) of potential reviewers and let me know if there's anyone who you think would be especially well suited. Thanks!
