### 2019-11-11 b9af5ef
Use PyPI sdist to build conda package
In addition, specify KDEpy dependency via my conda channel package.
