+++ conda/build.sh
@@ -1,10 +1,0 @@
-#
-# build.sh
-# Install script for conda package.
-#
-# Daniel Mentiplay, 2019
-#
-
-#--- Plonk Python package
-
-$PYTHON setup.py install
