+++ setup.py
@@ -47,4 +47,13 @@ setup(
     packages=packages,
     license='MIT',
     install_requires=install_requires,
+    classifiers=[
+        "Development Status :: 5 - Production/Stable",
+        "License :: OSI Approved :: MIT License",
+        "Programming Language :: Python :: 3.6",
+        "Programming Language :: Python :: 3.7",
+        "Operating System :: OS Independent",
+        "Topic :: Scientific/Engineering :: Astronomy",
+        "Topic :: Scientific/Engineering :: Visualization",
+    ],
 )
