### 2019-11-11 a0dd8a6
Add pip installation instructions to README
