### 2019-11-11 #1876 @dmentipl
Hi @dfm, thanks for agreeing to review my submission!

I had a look at the list, and no one in particular stood out. I suppose anyone with Python and astrophysics/astronomy expertise would be suitable.
