### 2019-11-11 8cdc31a
Add function to convert arrays to pandas DataFrame
