### 2019-11-11 c4956bb
Satisfy mypy type checking
