### 2019-11-11 e27feb2
Add API to documentation
