+++ plonk/visualize/plot.py
@@ -37,9 +37,9 @@ def plot(
 
     Parameters
     ----------
-    scalar_data, optional
+    scalar_data
         The 1d array (N,) of scalar data to visualize.
-    vector_data, optional
+    vector_data
         The 2d array (N, 2) of vector data to visualize.
     x_coordinate
         The x-position on the particles, where x is the required plot
@@ -47,7 +47,7 @@ def plot(
     y_coordinate
         The y-position on the particles, where y is the required plot
         y-axis.
-    z_coordinate, optional
+    z_coordinate
         The z-position on the particles, where z is the depth-axis for
         the required plot.
     extent
@@ -56,13 +56,13 @@ def plot(
         The particle mass for each particle.
     smoothing_length
         The smoothing length for each particle.
-    axis, optional
+    axis
         A matplotlib axis handle.
-    scalar_options, optional
+    scalar_options
         A dictionary of options for scalar plots.
-    vector_options, optional
+    vector_options
         A dictionary of options for vector plots.
-    interpolation_options, optional
+    interpolation_options
         A dictionary of options for interpolation.
     """
     return Visualization().plot(
