+++ docs/source/plonk.visualize.rst
@@ -1,0 +1,38 @@
+plonk.visualize package
+=======================
+
+Submodules
+----------
+
+plonk.visualize.interpolation module
+------------------------------------
+
+.. automodule:: plonk.visualize.interpolation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+plonk.visualize.plot module
+---------------------------
+
+.. automodule:: plonk.visualize.plot
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+plonk.visualize.visualization module
+------------------------------------
+
+.. automodule:: plonk.visualize.visualization
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: plonk.visualize
+   :members:
+   :undoc-members:
+   :show-inheritance:

[…]
