### 2019-11-11 a2cfb8e
Add readthedocs configuration
