+++ docs/environment.yml
@@ -1,0 +1,16 @@
+channels:
+  - conda-forge
+  - dmentipl
+
+dependencies:
+  - python>=3.6
+  - astropy
+  - h5py
+  - KDEpy
+  - matplotlib
+  - numpy
+  - pandas
+  - pint
+  - scikit-image>=0.16
+  - scipy
+  - sphinx
