### 2019-11-11 #1884 @whedon
[REVIEW]: Plonk: Smoothed particle hydrodynamics analysis and visualization with Python
**Submitting author:** @dmentipl (<a href="http://orcid.org/0000-0002-5526-8798">Daniel Mentiplay</a>)
**Repository:** <a href="https://github.com/dmentipl/plonk" target ="_blank">https://github.com/dmentipl/plonk</a>
**Version:** 0.2.1
