### 2019-11-11 dmentipl/plonk#2 @zingale
[JOSS review] sample data needed
### Is your feature request related to a problem? Please describe

The docs have a getting started section, but in order to play along, we need the sample data.  The results of each step should be provided as well so we can verify the library is working as intended.

### Describe the solution you'd like

sample dataset referenced in the docs should be included

### Describe alternatives you've considered

A clear and concise description of any alternative solutions or features you've considered.

### Additional context

Add any other context or screenshots about the feature request here.
