### 2019-11-11 dmentipl/plonk#2 @zingale
https://github.com/openjournals/joss-reviews/issues/1884
