### 2019-11-12 dmentipl/plonk#2 @dmentipl
Hello @zingale, I've uploaded some sample [Phantom](https://phantomsph.bitbucket.io/) data at: https://anaconda.org/dmentipl/example-phantom-data/files.

The data is from a simulation of a gas accretion disc with two sink particles, one representing the central star, and the other a planet. It includes:

- two simulation snapshots as HDF5 files,
- one text file with globally-averaged quantities,
- two text files with sink particle quantities.

I have not yet done, but will add output to the documentation showing results of each step.
