### 2019-11-21 68592f6
Expose Profile in __init__
