+++ plonk/__init__.py
@@ -66,13 +66,14 @@ import pint
 units = pint.UnitRegistry()
 
 from . import analysis, simulation, snap, utils, visualize
+from .analysis import Profile
 from .simulation import Evolution, Simulation, load_ev, load_sim
 from .snap import Snap, load_snap
 from .visualize import Visualization
 
 
 __all__ = (
-    ['Evolution', 'Simulation', 'Snap', 'Visualization']  # Classes
+    ['Evolution', 'Profile', 'Simulation', 'Snap', 'Visualization']  # Classes
     + ['analysis', 'simulation', 'snap', 'utils', 'visualize']  # Packages
     + ['load_snap', 'load_ev', 'load_sim']  # User functions
     + ['units']
