### 2019-11-21 d810cc5
Add JOSS badge to README
