### 2019-11-21 ccc5f51
Add to documentation
Add the following:
- Add to the "getting started guide" and renamed "overview".
- Add a quickstart guide for typical tasks, without much explanation.
- Tidy up the API documentation.
- Add mentions of other useful software.
- Add version string.
