+++ docs/source/getting-help.rst
@@ -1,0 +1,16 @@
+============
+Getting help
+============
+
+If you need help, try the following, in order:
+
+1. Check the documentation.
+2. Ask questions on Stack Overflow using the
+   `[plonk] <https://stackoverflow.com/questions/tagged/plonk>`_ tag.
+3. File an issue, as a
+   `bug report <https://github.com/dmentipl/plonk/issues/new?assignees=&labels=&template=bug_report.md&title=>`_
+   or `feature request <https://github.com/dmentipl/plonk/issues/new?assignees=&labels=&template=feature_request.md&title=>`_,
+   using the issue tracker on GitHub.
+
+If you don't get an immediate response, please be patient. Plonk is maintained
+by one person, `@dmentipl <https://github.com/dmentipl>`_.
