+++ docs/source/plonk.utils.rst
@@ -1,30 +1,0 @@
-plonk.utils package
-===================
-
-Submodules
-----------
-
-plonk.utils.exceptions module
------------------------------
-
-.. automodule:: plonk.utils.exceptions
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-plonk.utils.geometry module
----------------------------
-
-.. automodule:: plonk.utils.geometry
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: plonk.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
