+++ docs/source/plonk.simulation.rst
@@ -1,30 +1,0 @@
-plonk.simulation package
-========================
-
-Submodules
-----------
-
-plonk.simulation.evolution module
----------------------------------
-
-.. automodule:: plonk.simulation.evolution
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-plonk.simulation.simulation module
-----------------------------------
-
-.. automodule:: plonk.simulation.simulation
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: plonk.simulation
-   :members:
-   :undoc-members:
-   :show-inheritance:

[…]
