### 2019-11-21 dmentipl/plonk#2 @dmentipl
Hi @zingale, sorry for the slow response.

I've added to the documentation. And put up a more extensive data set to test at: https://anaconda.org/dmentipl/plonk_example_data.
