### 2019-11-21 9fc94f0
Fix bug in applying rotations to snapshots
