### 2019-11-22 9c1236e
Fix for accessing derived arrays with SubSnap
Without this fix, you cannot access derived arrays in a SubSnap due to
the way __getitem__ works.
