### 2019-11-22 1aa59c8
Fix type annotation in geometry.py
