+++ docs/source/examples.rst
@@ -1,0 +1,239 @@
+========
+Examples
+========
+
+---------------
+Rotate snapshot
+---------------
+
+Rotate a snapshot and plot column density.
+
+.. code-block:: pycon
+
+    >>> import matplotlib.pyplot as plt
+    >>> import numpy as np
+    >>> import plonk
+    >>> from scipy.spatial.transform import Rotation
+
+    # Load the snapshot
+    >>> snap = plonk.load_snap('disc_00030.h5')
+
+    # Define a rotation vector via scipy.spatial.transform
+    >>> rotation_angle = np.pi / 2.5
+    >>> rotation_vector = np.array([1, 1, 0])
+    >>> rotation_vector = rotation_vector / np.linalg.norm(rotation_vector)
+    >>> rotation = Rotation.from_rotvec(rotation_angle * rotation_vector)
+
+    # Apply the rotation to the snapshot
+    >>> snap.rotate(rotation)
+
+    # Plot
+    >>> plonk.visualize.render(snap, quantity='rho')
+    >>> plt.show()
+
+.. figure:: _static/rotate.png
+
+-------------
+Cross section
+-------------
+
+Plot cross section at z=0.
+
+.. code-block:: pycon
+
+    >>> import matplotlib.pyplot as plt
+    >>> import plonk
+
+    # Load the snapshot
+    >>> snap = plonk.load_snap('disc_00030.h5')
+
+    # Plot cross section
+    >>> interpolation_options = {'cross_section': 0.0}
+    >>> plonk.visualize.render(
+    ...     snap,
+    ...     quantity='rho',
+    ...     interpolation_options=interpolation_options,
+    ...     extent=(-150, 150, -150, 150),
+    ... )
+    >>> plt.show()
+
+.. figure:: _static/cross_section.png
+
+------------
+Side by side
+------------
+
+Plot dust and gas side-by-side.
+
+.. code-block:: pycon
+
+    >>> import matplotlib.pyplot as plt
+    >>> import plonk
+
+    # Load the snapshot
+    >>> snap = plonk.load_snap('disc_00030.h5')
+
+    # Specify dust and gas subsnaps
+    >>> gas = snap[snap['dust_id'] == 0]
+    >>> dust = snap[snap['dust_id'] == 1]
+    >>> extent = (-150, 150, -150, 150)
+
+    # Make plot
+    >>> fig, axes = plt.subplots(ncols=2, sharey=True, figsize=(12, 5))
+    >>> plonk.visualize.render(
+    ...     gas,
+    ...     quantity='rho',
+    ...     extent=extent,
+    ...     scalar_options={'cmap': 'Blues_r'},
+    ...     axis=axes[0],
+    ... )
+    >>> plonk.visualize.render(
+    ...     dust,
+    ...     quantity='rho',
+    ...     extent=extent,
+    ...     scalar_options={'cmap': 'Reds_r'},
+    ...     axis=axes[0],
+    ... )
+    >>> plt.show()
+
+.. figure:: _static/dust_and_gas.png
+
+------------------------------------
+Vector plot with coordinate tranform
+------------------------------------
+
+Transform to cylindrical coordinates and plot velocity vectors.
+
+.. code-block:: pycon
+
+    >>> import matplotlib.pyplot as plt
+    >>> import numpy as np
+    >>> import plonk
+
+    # Load snapshot
+    >>> snap = plonk.load_snap('disc_00030.h5')
+
+    # Transform to cylindrical coordinates (r, phi, z)
+    >>> plonk.utils.coordinate_transform(
+    ...     position=snap['position'],
+    ...     velocity=snap['velocity'],
+    ...     geometry_from='cartesian',
[…]
