### 2019-11-22 #1884 @dmentipl
Hi @zingale and @matthewturk. I've added to the documentation and provided a test data set at https://anaconda.org/dmentipl/plonk_example_data/.
