### 2019-11-23 dmentipl/plonk#2 @zingale
thanks, that looks great.  I'll try it out this weekend and then close this issue.
