### 2019-11-23 #1884 @zingale
My main comment was address and I was able to follow the analysis to do the example analysis and visualization.  I am happy to accept.
