### 2019-11-25 #1884 @dfm
@dmentipl: Looking good! Can you make the following small changes:

- Update the scikit-image reference to the preferred citation as described here: https://scikit-image.org/
- Update the scipy and numpy references as requested here: https://www.scipy.org/citing.html
- Add the DOI for the Jupyter reference: [10.3233/978-1-61499-649-1-87](http://ebooks.iospress.nl/publication/42900)
- Add a capitalization guard to the title of {{Python and HDF5}}
- ASCL references don't render well in TeX so please add `howpublished = {Astrophysics Source Code Library}` to the entry to make it clearer
- In paragraph 3, "...a set of particles with mass" should be something like "...a set of massive particles".
- At the bottom of page 1, the sentence starting with "Tasks such as:..." is not complete. It could be combined with the previous sentence or expanded into a full sentence of its own right.
- On page 2, the sentence about yt should include more information to draw comparisons with Plonk. For example, my understanding is that yt was originally designed for grid codes with SPH support added more recently. If that's true, perhaps that sentence could include a comment about that.
- The last sentence of the main body could use more detail. Would users of Phantom SPH find Plonk useful because that's what it was originally designed to visualize? If so, that would be useful to include.
