### 2019-11-26 9b7a943
Add requested changes to paper
See the following issue comment:
https://github.com/openjournals/joss-reviews/issues/1884#issuecomment-558133007
