+++ docs/paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 
 Hydrodynamical simulations are a standard part of the astronomer’s toolkit. The knowledge required to run these simulations and analyze the output can be specialized. Many of the codes written to run and analyze these simulations are written in compiled languages such as Fortran and C++. The advantage of using compiled languages is performance. One disadvantage is that expertise in these languages is becoming less widely available than other, typically interpreted, languages such as Python. Access to analysis and visualization of hydrodynamical simulations should be available to a wide audience. Much of the modern astronomy pipeline is written in Python.
 
-Smoothed particle hydrodynamics (SPH) is a method for solving the equations of hydrodynamics [@Lucy:1977; @Gingold:1977; @Price:2012]. The fluid is discretized as a set of particles with mass. The density field is computed as a sum over particles weighted by a smoothing kernel. The particles are evolved by solving the equations of hydrodynamics written in Lagrangian form.
+Smoothed particle hydrodynamics (SPH) is a method for solving the equations of hydrodynamics [@Lucy:1977; @Gingold:1977; @Price:2012]. The fluid is discretized as a set of massive particles. The density field is computed as a sum over particles weighted by a smoothing kernel. The particles are evolved by solving the equations of hydrodynamics written in Lagrangian form.
 
 Visualization of SPH data is complicated by the fact that the data is represented on a set of particles arranged irregularly in space, as opposed to a regular grid. One visualization method is to plot the particles as a scatter plot, possibly with a color to represent some quantity defined on the particles. An alternative is to interpolate quantities defined on the particles to a two-dimensional pixel array. The latter approach provides an accurate representation of the data, in that it uses the same underlying approach to computing density as in the SPH method.
 
@@ -30,16 +30,16 @@ SPH simulations typically output files which contain quantities defined on the p
 
 ``Plonk`` can read data from these various data sets and provide access to the underlying data as ``NumPy`` arrays. Large arrays are loaded only as required. The entire data set does not need to be read into memory at once. After the array is loaded from file it is cached in memory for later use. ``Plonk`` can read particles representing gas, dust (including multiple species), or other fluids, as well as sink particles (representing compact objects like stars or planets).
 
-``Plonk`` has tools to perform typical tasks associated with the analysis and visualization of smoothed particle hydrodynamics data. Tasks such as: calculating extra quantities on the particles, calculating and plotting radial profiles, and visualization of any quantity defined on the particles via kernel density estimation. The interpolation component of the visualization package makes use of ``KDEpy``, a Python package for weighted kernel density estimation using fast Fourier transforms [@Odland:2018]. As such it has performance on par with other (compiled) SPH visualization programs such as Splash [@Price:2007].
+``Plonk`` has tools to perform typical tasks associated with the analysis and visualization of smoothed particle hydrodynamics data, such as: calculating extra quantities on the particles, calculating and plotting radial profiles, and visualization of any quantity defined on the particles via kernel density estimation. The interpolation component of the visualization package makes use of ``KDEpy``, a Python package for weighted kernel density estimation using fast Fourier transforms [@Odland:2018]. As such it has performance on par with other (compiled) SPH visualization programs such as Splash [@Price:2007].
 
[…]
