+++ docs/paper/paper.bib
@@ -13,7 +13,7 @@
   adsurl = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A}
 }
 
-@misc{Benitez-Llambay:2015,
+@MISC{Benitez-Llambay:2015,
   author = {Alejandro Benitez-Llambay},
   title = {py-sphviewer: Py-SPHViewer v1.0.0},
   month = jul,
@@ -26,7 +26,7 @@
   keywords = {python, hdf5},
   year = {2013},
   publisher = {O'Reilly},
-  title = {Python and HDF5},
+  title = {{Python and HDF5}},
   author = {Andrew Collette}
 }
 
@@ -66,7 +66,8 @@
   Organization = {IOS Press},
   Pages = {87 - 90},
   Title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows},
-  Year = {2016}
+  Year = {2016},
+  doi = {10.3233/978-1-61499-649-1-87}
 }
 
 @ARTICLE{Lucy:1977,
@@ -103,23 +104,15 @@
   url = {https://doi.org/10.5281/zenodo.2392268}
 }
 
-@ARTICLE{Oliphant:2007,
-  abstract = {Not Available},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System},
-  adsurl = {https://ui.adsabs.harvard.edu/abs/2007CSE.....9c..10O},
-  author = {{Oliphant}, T.~E.},
-  doi = {10.1109/MCSE.2007.58},
-  journal = {Computing in Science and Engineering},
-  number = {3},
-  pages = {10-20},
-  title = {{Python for Scientific Computing}},
-  volume = {9},
-  year = {2007}
+@BOOK{Oliphant:2006,
+  year = {2006},
+  publisher = {USA: Trelgol Publishing},
+  title = {A guide to NumPy},
+  author = {Travis E. Oliphant}
 }
 
 @MISC{Pontzen:2013,
-       author = {{Pontzen}, Andrew and {Ro{\v{s}}kar}, Rok and {Stinson}, Greg and
-         {Woods}, Rory},
+       author = {{Pontzen}, Andrew and {Ro{\v{s}}kar}, Rok and {Stinson}, Greg and {Woods}, Rory},
         title = "{pynbody: N-Body/SPH analysis for python}",
      keywords = {Software},
          year = "2013",
@@ -129,7 +122,8 @@
 archivePrefix = {ascl},
        eprint = {1305.002},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2013ascl.soft05002P},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+ howpublished = {Astrophysics Source Code Library}
 }
 
 @ARTICLE{Price:2007,
@@ -186,26 +180,8 @@ archivePrefix = {ascl},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{van-der-Walt:2014,
-  author = {{van der Walt}, Stefan and {Sch{\"o}nberger}, Johannes L. and {Nunez-Iglesias}, Juan and {Boulogne}, Fran{\c{c}}ois and {Warner}, Joshua D. and {Yager}, Neil and {Gouillart}, Emmanuelle and {Yu}, Tony and {scikit-image contributors}, the},
-  title = "{scikit-image: Image processing in Python}",
-  journal = {arXiv e-prints},
-  keywords = {Computer Science - Mathematical Software, Computer Science - Computer Vision and Pattern Recognition},
-  year = "2014",
-  month = "Jul",
-  eid = {arXiv:1407.6245},
-  pages = {arXiv:1407.6245},
[…]
