### 2019-11-26 #1884 @dmentipl
Hi @dfm, I've made all changes requested. See this [commit](https://github.com/dmentipl/plonk/commit/9b7a943e9fcd4bfc506ca5be726a4ba14b7fdb6c).

Hopefully the changes I've made to the last paragraph in the main body cover the last two points above.

@matthewturk I hope my changes convey the state of yt with regards to the demeshening.
