### 2019-11-26 #1884 @dfm
@dmentipl: Looks good to me! Can you update generate a new Zenodo archive with a title and author list that match the manuscript? Once you do that, report the new DOI here.
