### 2019-11-26 cf1f92b
Add doc8 to dev environment.yml
