### 2019-11-26 1c639a7
Set visualize.render quantity directly
We attempt to load `snap[quantity]` and catch exceptions. Also, by
default plot all particles, not those within the 99th percentile. To
change the extent use the `extent` argument.
