### 2019-11-26 b1fd3cd
Fix doc8 warnings
