+++ plonk/visualize/plot.py
@@ -2,7 +2,6 @@
 
 from typing import Any, Dict, Optional, Tuple
 
-import numpy as np
 from numpy import ndarray
 
 from ..snap.snap import Snap
