### 2019-11-26 d5d4ad1
Add small documentation fixes
