+++ plonk/__init__.py
@@ -80,4 +80,4 @@ __all__ = (
 )
 
 # Canonical version number
-__version__ = '0.2.0'
+__version__ = '0.2.1'
