### 2019-11-27 #1884 @matthewturk
Hi!  For what it's worth, I definitely think it's appropriate for me not to
be an author.

On Tue, Nov 26, 2019 at 7:12 PM Daniel Mentiplay <notifications@github.com>
wrote:

> @dfm <https://github.com/dfm>: The Zenodo archive DOI is
> 10.5281/zenodo.3554568
>
> Also, I updated the Plonk version to v0.2.1.
>
> The Zenodo archive automatically generated from the GitHub release
> included Matthew as an author, as he contributed a pull request adding a
> reference to yt in the manuscript. However, I removed him from the Zenodo
> archive author list. I hope that's appropriate?
>
