### 2019-11-27 d7d1fa0
Fix version for building Conda package
