### 2019-11-27 92d4aea
Add PyPI and Conda instructions to CONTRIBUTING.md
