### 2019-11-27 #1884 @dfm
This looks good to go for me! Pinging @openjournals/joss-eics for final processing.
