### 2019-11-27 #1884 @dmentipl
@dfm The proof looks good to me. Let me know if there's anything left to do.
