### 2019-12-01 #1884 @Kevin-Mattheus-Moerman
I checked the paper and Zenodo archive as well and all looks good.
