### 2019-12-01 #1884 @dfm
@dmentipl: congrats - your paper is now published! 🎉

@zingale, @matthewturk: thanks 💯 for your constructive reviews!
