### 2019-12-01 dd8f967
Use JOSS for citation, remove ASCL citation
