+++ CITATION.bib
@@ -1,13 +1,11 @@
-@MISC{2019ascl.soft07009M,
+@ARTICLE{2019JOSS....4.1884M,
        author = {{Mentiplay}, Daniel},
-        title = "{Plonk: Smoothed particle hydrodynamics data analysis and visualization}",
-     keywords = {Software},
+        title = "{Plonk: Smoothed particle hydrodynamics analysis and visualization with Python}",
+      journal = {The Journal of Open Source Software},
          year = "2019",
-        month = "Jul",
-          eid = {ascl:1907.009},
-        pages = {ascl:1907.009},
-archivePrefix = {ascl},
-       eprint = {1907.009},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2019ascl.soft07009M},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+        month = "Dec",
+       volume = {4},
+       number = {44},
+        pages = {1884},
+          doi = {10.21105/joss.01884},
 }
