### 2019-12-01 9cf6b5a
Change citation to JOSS paper in docs
