### 2019-12-02 44db736
Add analysis quantities
Also, remove `mask` keyword argument. This is not required as we can
pass in a SubSnap which is already "masked".
