### 2019-12-02 #1884 @Kevin-Mattheus-Moerman
@openjournals/dev this DOI is not resolving yet, can you check?
