### 2019-12-02 #1884 @danielskatz
The DOI resolves for me, and the paper looks fine.
